Feature: add a JWT claim (add an admin status to the token)

+++ resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import get_jwt_claims, jwt_required
 from flask_restful import Resource, reqparse
 
 from models.item import ItemModel
@@ -42,7 +42,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': "Admin privilege required."}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
